See #42.  Refactored.

+++ src/ccaf_housekeeping_python_lib/drop_table_with_associated_resources.py
@@ -1,5 +1,6 @@
+from typing import Tuple
 from cc_clients_python_lib.flink_client import FlinkClient
-from cc_clients_python_lib.common import HttpStatus
+from cc_clients_python_lib.http_status import HttpStatus
 
 
 __copyright__  = "Copyright (c) 2025 Jeffrey Jonathan Jennings"
@@ -12,17 +13,21 @@ __status__     = "dev"
 
 class DropTableWithAssociatedResources():
     def __init__(self, flink_config: dict):
-        # Instantiate the Flinklient classs.
+        # Instantiate the Flink client class.
         self.flink_client = FlinkClient(flink_config)
 
-    def drop_table(self, catalog_name: str, database_name: str, table_name: str):
-        http_status_code, error_message, response = self.flink_client.get_compute_pool_list()
+    def drop_table(self, catalog_name: str, database_name: str, table_name: str) -> Tuple[bool, str]:
+        """Drop the table and associated resources."""
+        # Get the compute pool.
+        http_status_code, error_message, response = self.flink_client.get_compute_pool()
+        if http_status_code != HttpStatus.OK:
+            return False, error_message
 
-        for item in response.get("data"):
-            logger.info("%s, %d, %d, %s", item.get("id"), item.get("status").get("current_cfu"), item.get("spec").get("max_cfu"), item.get("status").get("phase"))
-        
-
-        http_status_code, error_message, response = self.flink_client.get_statement_list()
+        # Get the statement list.
+        max_cfu = response.get("spec").get("max_cfu")
+        http_status_code, error_message, response = self.flink_client.get_statement_list(max_cfu)
+        if http_status_code != HttpStatus.OK:
+            return False, error_message
 
         http_status_code, error_message, response = self.flink_client.submit_statement(f"drop-{table_name}",
                                                                                        f"DROP TABLE IF EXISTS {table_name};",
